requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu 
torchvision==0.8.2+cpu
faiss-cpu
streamlit
sentence_transformers
